refactor: extract plan-ready messaging actions into app hook
